Fixed String to BigDecimal TypeConverter; no longer causes number format exception. Can now successfully add and update to Room DB

+++ app/src/main/java/edu/msoe/receiptstorageapp/Receipt.kt
@@ -11,7 +11,7 @@ data class Receipt (
     @PrimaryKey val receiptId: UUID,
     val date: Date,
     val vendorName: String,
-    val grandTotal: Double,
+    val grandTotal: BigDecimal,
 //   TODO: Add @ForeignKey val imageSetId: UUID,
 //    then need DAO operations for updating imageSets/images
 )
